+++ modules/apps/firefox.nix
@@ -30,9 +30,9 @@ in {
           "layout.spellcheckDefault" = 2;
 
           # New Tab
-          "browser.newtabpage.activity-stream.feeds.section.highlights" = true;
+          "browser.newtabpage.activity-stream.feeds.section.highlights" = false;
           "browser.newtabpage.activity-stream.feeds.section.topstories" = false;
-          "browser.newtabpage.activity-stream.feeds.topsites" = true;
+          "browser.newtabpage.activity-stream.feeds.topsites" = false;
           "browser.newtabpage.activity-stream.showSearch" = false;
           "browser.newtabpage.activity-stream.showSponsored" = false;
           "browser.newtabpage.activity-stream.showSponsoredTopSites" = false;