+++ modules/shell/zsh.nix
@@ -46,13 +46,13 @@ in {
               bindkey '^[[F'    end-of-line
               bindkey '^[[1;5C' forward-word
               bindkey '^[[1;5D' backward-word
+              bindkey '^[[D'    backward-char
+              bindkey '^[[C'    forward-char
+              bindkey '^[[A'    up-line-or-search
+              bindkey '^[[B'    down-line-or-search
               bindkey '^H'      backward-kill-word
               bindkey '^[[3;5~' kill-word
               bindkey '^J'      backward-kill-line
-              bindkey '^[[D'    backward-char
-              bindkey '^[[C'    forward-char
-              bindkey '^[[A'    up-line-or-beginning-search
-              bindkey '^[[B'    down-line-or-beginning-search
             '';
 
           # Command History